# Imtihon 2/api_gataway/docs/swagger.yaml
definitions:
...
  genproto.CreateProjectRequest:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  genproto.CreateSkillRequest:
    properties:
      id:
        type: string
      level:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
  genproto.Skill:
    properties:
      id:
        type: string
      level:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  genproto.Skills:
    properties:
      Skills:
        items:
          $ref: '#/definitions/genproto.Skill'
        type: array
    type: object
...
  genproto.UpdateProjectRequest:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  genproto.UpdateSkillRequest:
    properties:
      id:
        type: string
      level:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
  title: Portfolio swagger UI
paths:
...
  /project/create:
    post:
      consumes:
      - application/json
      description: This API for creating project
      parameters:
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Create Project
      tags:
      - project
  /project/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting project
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Delete Project
      tags:
      - project
  /project/get/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting project
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Project
      tags:
      - project
  /project/getall:
    get:
      consumes:
      - application/json
      description: This API for getting projects
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Project url
        in: query
        name: url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Projects
      tags:
      - project
  /project/getallbyuser:
    get:
      consumes:
      - application/json
      description: This API for getting projects
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Projects By User
      tags:
      - project
  /project/update:
    put:
      consumes:
      - application/json
      description: This API for updating project
      parameters:
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Update Project
      tags:
      - project
  /skill/create:
    post:
      consumes:
      - application/json
      description: This API for creating skill
      parameters:
      - description: Skill
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateSkillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Create Skill
      tags:
      - skill
  /skill/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting skill
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Delete Skill
      tags:
      - skill
  /skill/get/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting skill
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Skill
      tags:
      - skill
  /skill/getall:
    get:
      consumes:
      - application/json
      description: This API for getting skills
      parameters:
      - description: Skill Name
        in: query
        name: name
        type: string
      - description: Skill Level
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Skills'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Skills
      tags:
      - skill
  /skill/getallbyuser:
    get:
      consumes:
      - application/json
      description: This API for getting skills
      parameters:
      - description: Skill User Id
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Skills'
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get Skills By User
      tags:
      - skill
  /skill/update:
    put:
      consumes:
      - application/json
      description: This API for updating skill
      parameters:
      - description: Skill
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateSkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Update Skill
      tags:
      - skill
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"